backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Model Inference
faster-whisper==1.0.0
llama-cpp-python==0.2.56
piper-tts==1.2.0

# Audio Processing
numpy==1.26.3
soundfile==0.12.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
tqdm==4.66.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
